conditional rm

+++ tpu_config_test.py
@@ -4,7 +4,9 @@ print("Tensorflow version " + tf.__version__)
 
 os.system("export TPU_NAME=tpu-vm-1")
 os.system("export TPU_LOAD_LIBRARY=0")
-os.system("sudo rm -r /tmp/") # removing lockfile that seems to prevent tpu server startup
+
+if os.path.exists("/tmp/"):
+    os.system("sudo rm -r /tmp/") # removing lockfile that seems to prevent tpu server startup
 
 cluster_resolver = tf.distribute.cluster_resolver.TPUClusterResolver()
 print('Running on TPU ', cluster_resolver.cluster_spec().as_dict()['worker'])
